+++ sea-orm-codegen/src/entity/transformer.rs
@@ -1,5 +1,6 @@
 use crate::{
-    Column, ConjunctRelation, Entity, EntityWriter, Error, PrimaryKey, Relation, RelationType,
+    ActiveEnum, Column, ConjunctRelation, Entity, EntityWriter, Error, PrimaryKey, Relation,
+    RelationType,
 };
 use sea_query::TableStatement;
 use std::collections::HashMap;
@@ -9,6 +10,7 @@ pub struct EntityTransformer;
 
 impl EntityTransformer {
     pub fn transform(table_stmts: Vec<TableStatement>) -> Result<EntityWriter, Error> {
+        let mut enums: HashMap<String, ActiveEnum> = HashMap::new();
         let mut inverse_relations: HashMap<String, Vec<Relation>> = HashMap::new();
         let mut conjunct_relations: HashMap<String, Vec<ConjunctRelation>> = HashMap::new();
         let mut entities = HashMap::new();
@@ -22,7 +24,15 @@ impl EntityTransformer {
                 }
             };
             let table_name = match table_create.get_table_name() {
-                Some(s) => s,
+                Some(table_ref) => match table_ref {
+                    sea_query::TableRef::Table(t)
+                    | sea_query::TableRef::SchemaTable(_, t)
+                    | sea_query::TableRef::DatabaseSchemaTable(_, _, t)
+                    | sea_query::TableRef::TableAlias(t, _)
+                    | sea_query::TableRef::SchemaTableAlias(_, t, _)
+                    | sea_query::TableRef::DatabaseSchemaTableAlias(_, _, t, _) => t.to_string(),
+                    _ => unimplemented!(),
+                },
                 None => {
                     return Err(Error::TransformError(
                         "Table name should not be empty".into(),
@@ -44,12 +54,51 @@ impl EntityTransformer {
                         > 0;
                     col
                 })
+                .map(|col| {
+                    if let sea_query::ColumnType::Enum(enum_name, values) = &col.col_type {
+                        enums.insert(
+                            enum_name.clone(),
+                            ActiveEnum {
+                                enum_name: enum_name.clone(),
+                                values: values.clone(),
+                            },
+                        );
+                    }
+                    col
+                })
                 .collect();
-            let relations = table_create
+            let mut ref_table_counts: HashMap<String, usize> = HashMap::new();
+            let relations: Vec<Relation> = table_create
                 .get_foreign_key_create_stmts()
                 .iter()
                 .map(|fk_create_stmt| fk_create_stmt.get_foreign_key())
-                .map(|tbl_fk| tbl_fk.into());
+                .map(|tbl_fk| {
+                    let ref_tbl = tbl_fk.get_ref_table().unwrap();
+                    if let Some(count) = ref_table_counts.get_mut(&ref_tbl) {
+                        if *count == 0 {
+                            *count = 1;
+                        }
+                        *count += 1;
+                    } else {
+                        ref_table_counts.insert(ref_tbl, 0);
+                    };
+                    tbl_fk.into()
+                })
+                .collect::<Vec<_>>()
+                .into_iter()
+                .rev()
+                .map(|mut rel: Relation| {
+                    rel.self_referencing = rel.ref_table == table_name;
+                    if let Some(count) = ref_table_counts.get_mut(&rel.ref_table) {
+                        rel.num_suffix = *count;
+                        if *count > 0 {
+                            *count -= 1;
+                        }
+                    }
+                    rel
+                })
+                .rev()
+                .collect();
             let primary_keys = table_create
                 .get_indexes()
                 .iter()
@@ -67,12 +116,21 @@ impl EntityTransformer {
             let entity = Entity {
                 table_name: table_name.clone(),
                 columns,
-                relations: relations.clone().collect(),
+                relations: relations.clone(),
                 conjunct_relations: vec![],
                 primary_keys,
             };
             entities.insert(table_name.clone(), entity.clone());
             for (i, mut rel) in relations.into_iter().enumerate() {
+                // This will produce a duplicated relation
+                if rel.self_referencing {
+                    continue;
+                }
+                // This will cause compile error on the many side,
+                // got relation variant but without Related<T> implemented
+                if rel.num_suffix > 0 {
+                    continue;
+                }
                 let is_conjunct_relation = entity.primary_keys.len() == entity.columns.len()
                     && rel.columns.len() == 2
                     && rel.ref_columns.len() == 2
@@ -134,6 +192,7 @@ impl EntityTransformer {
         }
         Ok(EntityWriter {
             entities: entities.into_iter().map(|(_, v)| v).collect(),
+            enums,
         })
     }
 }