+++ tests/common/features/active_enum_child.rs
@@ -3,10 +3,11 @@ use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
 #[cfg_attr(feature = "sqlx-postgres", sea_orm(schema_name = "public"))]
-#[sea_orm(table_name = "active_enum")]
+#[sea_orm(table_name = "active_enum_child")]
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
+    pub parent_id: i32,
     pub category: Option<Category>,
     pub color: Option<Color>,
     pub tea: Option<Tea>,
@@ -14,25 +15,29 @@ pub struct Model {
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
-    #[sea_orm(has_many = "super::active_enum_child::Entity")]
-    ActiveEnumChild,
+    #[sea_orm(
+        belongs_to = "super::active_enum::Entity",
+        from = "Column::ParentId",
+        to = "super::active_enum::Column::Id"
+    )]
+    ActiveEnum,
 }
 
-impl Related<super::active_enum_child::Entity> for Entity {
+impl Related<super::active_enum::Entity> for Entity {
     fn to() -> RelationDef {
-        Relation::ActiveEnumChild.def()
+        Relation::ActiveEnum.def()
     }
 }
 
-pub struct ActiveEnumChildLink;
+pub struct ActiveEnumLink;
 
-impl Linked for ActiveEnumChildLink {
+impl Linked for ActiveEnumLink {
     type FromEntity = Entity;
 
-    type ToEntity = super::active_enum_child::Entity;
+    type ToEntity = super::active_enum::Entity;
 
     fn link(&self) -> Vec<RelationDef> {
-        vec![Relation::ActiveEnumChild.def()]
+        vec![Relation::ActiveEnum.def()]
     }
 }